+++ src/html/public/js/types/adminlte.d.ts
@@ -1,3 +1,4 @@
+import { initialize, teardown } from './util/index.js';
 import Layout from './layout.js';
 import CardWidget from './card-widget.js';
 import Treeview from './treeview.js';
@@ -6,4 +7,4 @@ import FullScreen from './fullscreen.js';
 import PushMenu from './push-menu.js';
 import ColorMode from './color-mode.js';
 import { initAccessibility } from './accessibility.js';
-export { Layout, CardWidget, Treeview, DirectChat, FullScreen, PushMenu, ColorMode, initAccessibility };
+export { Layout, CardWidget, Treeview, DirectChat, FullScreen, PushMenu, ColorMode, initAccessibility, initialize, teardown };